Shifted enum values in ControlPrompts to start from 1L<<17 to avoid clash with DomainCategory
Accordingly had to change (int) to (long) in all places casting from/to ControlPrompts

+++ Src/CheckMade.Common.Model/Enums/EnumUiStringProvider.cs
@@ -58,7 +58,7 @@ public record EnumUiStringProvider
     }
     
     private void AddPrompt(ControlPrompts prompt, UiString uiString) =>
-        _promptsBuilder.Add(new EnumCallbackId((int)prompt), uiString);
+        _promptsBuilder.Add(new EnumCallbackId((long)prompt), uiString);
     
     private void AddCategory(DomainCategory category, UiString uiString) =>
         _categoryBuilder.Add(new EnumCallbackId((int)category), uiString);
